fix(engine): eviction-safe absolute cursor for FULL rejoin catch-up
FullHistory::snapshot_from indexed relative to the ring buffer's front,
so an eviction between two backpressured catch-up pumps silently skipped
exactly that many packets, producing a later spurious desync-kick in games
long enough to hit the 216k-packet cap. Track an absolute `evicted` counter
and feed the rejoiner by absolute sequence number; drop explicitly (with a
warning) if the needed packet was already evicted rather than skipping it.

- full_history: add `evicted` counter; first_seq/next_seq/snapshot_from_seq;
  remove the relative snapshot_from footgun.
- players: catchup_cursor u32 -> u64 (absolute).
- actions: init cursor to full_history.first_seq() on rejoin load.
- state: pump feeds by absolute seq, gap-drops a rejoiner that fell behind.
- tests: eviction-safety regression + seq unit tests.

+++ crates/spectre-engine/src/players.rs
@@ -60,7 +60,7 @@ pub struct Player {
     pub last_gproxy_wait_notice: Option<Instant>,
     pub gproxy_disconnect_notice_sent: bool,
     pub rejoin: RejoinStage,
-    pub catchup_cursor: Option<u32>,
+    pub catchup_cursor: Option<u64>,
     pub catching_up: bool,
 }
